ViewModel

+++ app/src/main/java/de/mlex/myquote2v2/data/Quote.kt
@@ -9,5 +9,5 @@ data class Quote(
     val author: String,
     val year: String,
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0
+    val id: Int = 0
 )
